root/worker_parser/requirements.txt
faststream[rabbit, cli]
selectolax
pydantic-settings
redis
httpx
fastapi
sqlalchemy[asyncio]
asyncpg